Improve herbivorous comportment
[EDIT:Experience]
Add possibility to do experience out of 100 range, which mean,
experience based on probabilities like 4/5 or 1/60 or that kind of
thing.

[EDIT:HexagonalMap]
Add border to animal display, to get more visibility on zoomed view.

[ADD/REFACTOR:probility.Operation]
Move the random between a and b function to a more dedicated file.

[EDIT:Group]
[EDIT:HerbivorousGroup]
[EDIT:AnimalState]
[ADD:HerbivorousState]
Improve herbivorous gesture:
- Now stopping on center of plant group, so group doesn't become a mess
on center.
- Add random group move to look for food, like a normal group would do.
- Add a speed parameter for each animal, based on the ability of specie,
but randomize to get different survival probability when being hunted.

[EDIT:SimulationConstant]
Reduce the plant spawn rate, and the animal individual random move rate.
Add the group move probability;